Cleaned up test issues.
